backend/requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.4.0
django-filter==25.2
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
